% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fbind.R
\name{fbind}
\alias{fbind}
\title{Concatenate factor levels}
\usage{
fbind(a, b)
}
\arguments{
\item{a}{a factor with at least one level}

\item{b}{a factor with at least one level}
}
\value{
a factor with combined levels
}
\description{
This function combines the levels of two factors into an integreated
set of levels.
}
\examples{

fbind(as.factor(c("dog", "cat")), as.factor(c("gerbil",   "parakeet")))





}
